+++ CHANGELOG.md
@@ -25,3 +25,13 @@
 
 ##  Version 0.0.7 (2023-05-07)
 - one more fix for langchain deepcopying and context sharing - fixing not propagating some additional info
+
+##  Version 0.0.8 (2023-05-10)
+- fix typing error in caching (caused problems with python 3.10+)
+- fixed nesting context problem
+- disabled exception on get cache that has not been initialized (reinitializing instead)
+
+## Version 0.1.0 (2023-05-15)
+- introduction of UnitTests
+- support for initialization of tracking_project from PROMPTWATCH_TRACKING_PROJECT Env variable
+- refactoring of the file structure